+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.0.8] - 2026-01-04
+
 ### Added
 - `Distinct<T>()` constraint for asserting pairwise distinctness of expressions
 - `RotateLeft()` and `RotateRight()` operations for bit-vectors